requirements.txt
# pip install -r requirements.txt

Mastodon.py
pandas
python-dotenv